https is required to access mediaDevices.
I figured out that in order to access the media devices (e.g., video and audio feeds, page must be served through https (i.e., http doesn't work).

Since the error message is not very helpful in this situation (i.e., it complains about how the qetUserMedia() method is not implemented on the device), I've added this commit / comment to remind myself in the future.