Refactor. Moved away BreathingSonarJS as a JS module.
Previously, I had defined BreathingSonarJS as a module to allow it to import the $1 recognizer library. However, I'm finding that using modules makes things more difficult to test locally, and interferes with default behavior in P5.js.

I still don't think this is the optimal solution, but I've simply inline the $1 recognizer tools into the BreathingSonarJS source. This allows BreathingSonarJS to be included in any html page using the standard <script> tag.