Store training examples to support exporting training data.
The Breathing Sonar tool should have the ability to export training examples to a file, so that they can then later be imported again. I've taken the first step in this commit by saving the training examples in an array which can be accessed and set by callers.

File format (.csv):
- no header row
- first column contains breathing pattern label
- all remaining columns contain signal time series

Note that these capabilities are illustrated in the index.js interactive demo - users press the space bar to log all training examples to an output file.