Refactor. train() now accepts optional `trainingData` parameter.
Previously, the train() method always used the current values in the rolling window as training input. This was fine for capturing training examples in real-time, but is inappropriate for loading previously recorded training input.

Refactored the method to optionally accept a `trainingData` parameter, which defaults to the current rolling window.