Added a cool down period between subsequent gesture recognitions.
Previously, it was possible for many occurrences of a gesture to be recognized on a single waveform (i.e., once per sampling frame). This was problematic because it could lead to registered callbacks being triggered over and over again.

I've fixed this issue for now by introducing a global "cool down period" between any and all recognized gestures. For now, this cool down is hard coded as the current rolling window length, and disables detection globally (i.e., the detection of any pattern disables detection of any other pattern for the duration of the cool down)