Moved to a 500ms rolling window to support more responsive game controls.
The current use case for the library is to support near real-time interaction with a side-scrolling game (i.e., cause the player to jump by taking a forceful inhale).

To support more responsive interactions, I've shortened the training / gesture recognition time window for now. I know that this probably isn't the best way to accomplish this for long term / flexible use of the library, but this will do for now. I've flagged this as an issue for future work on github.