Fix. Invert signal profile in interactive demo to reflect true signal readings.
Previously, I had decided to draw the breathing sonar signal upside-down in the interactive demo to be more intuitive for users (i.e., the signal actually decreases whenever breathing movement is detected).

In this commit, I've corrected this to avoid confusion between the visualization provided in the interactive demo vs. the signal readings that get logged to training data files.