Compute normalized sonar signal (i.e., [-1.0, 1.0]) based on min and max readings in rollingWindow.
The BreathingSonarJS library now computes a normalized version of the signal at each time step. This normalized reading is reported in the range [-1.0, 1.0], and is based on the current minimum and maximum readings found in the rollingWindow at any given time.

The formula used for this calculation is:

normalized = (reading - offset) / range

where...

  offset = (max + min) / 2
  range = max - min