Move away from gesture-recognition based approach.
So far, trying to identify stacks and PEP exhales as breathing "gestures" hasn't exactly been working so well. I'm going to abandon this idea for now in favor of a threshold-crossing approach (using the normalized signal, as well as derivative if necessary).

All of the gesture-related API tools are still available in the BreathingSonarJS library file, but have been removed from the interactive demo.