Fix. Invert Scrolling Line Graph output, and draw y-intercept.
I know this is about the third time I've made a commit regarding the sign of the sonar waveform drawn in the interactive demo, but I'm pretty sure I've got it correct this time... (I was getting mixed up regarding *increases* in y-values corresponding to *southward* translations on the canvas).

In this commit, I've corrected the waveform to correspond with the debugging signal output provided, and drawn in the y-intercept (i.e., y = 0) as a dotted line for reference.