Refactor. Incorporate highlight color definition into ScrollingLineGraph for portability.
Previously, the color being used to draw highlights in the ScrollingLineGraph class was defined as as global variable, which prevented easy portability of the ScrollingLineGraph class.

I've moved this color definition into the class for increased portability (i.e., defaults to pale green, but can be explicitly assigned in class constructor).